+++ src/const.ts
@@ -2,3 +2,4 @@ export const VALUE_TYPES_TO_SKIP = ["unknown", "array"];
 export const TABLE_NAME = "dataset";
 export const APIFY_FAVICON_URL = "https://apify.com/favicon.ico";
 export const APIFY_API_BASE_URL = "https://api.apify.com";
+export const APIFY_OPENAI_API_BASE_URL = "https://openrouter.apify.actor/api/v1";
