--- src/db.ts
@@ -1,5 +1,0 @@
-import { Database } from "bun:sqlite";
-
-export function createDatabase() {
-    return new Database(":memory:");
-}
